comit

+++ src/main/java/com/zr/test/demo/repository/SysUserDaoImpl.java
@@ -30,7 +30,7 @@ public class SysUserDaoImpl {
         return userDao.updateById(user);
     }
 
-    public int deleteById(Integer id){
+    public int deleteById(Long id){
         return userDao.deleteById(id);
     }
